fix(jota-memory): preserve BF16 fill values
Convert the supplied Number directly to float before BF16 encoding instead of narrowing it to a byte. Verify fractional, negative, and out-of-byte-range values across supported memory domains.

+++ jota/jota-memory/src/main/java/com/qxotic/jota/memory/MemoryHelpers.java
@@ -37,7 +37,7 @@ public final class MemoryHelpers {
             memoryOperations.fillShort(
                     memory, 0, byteSize, Float.floatToFloat16(value.floatValue()));
         } else if (dataType == DataType.BF16) {
-            memoryOperations.fillShort(memory, 0, byteSize, BFloat16.fromFloat(value.byteValue()));
+            memoryOperations.fillShort(memory, 0, byteSize, BFloat16.fromFloat(value.floatValue()));
         } else {
             throw new IllegalArgumentException("unsupported value " + value);
         }
